requirements.txt
absl-py==2.0.0
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
azure-cognitiveservices-speech==1.32.1
blis==0.7.11
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.0
click==7.1.2
cloudpathlib==0.16.0
confection==0.1.3
cymem==2.0.8
decorator==4.4.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl#sha256=6215d71a3212690e9aec49408a27e3fe6ad7cd6c715476e93d70dc784041e93e
exceptiongroup==1.1.3
ffmpeg-python==0.2.0
flatbuffers==23.5.26
frozenlist==1.4.0
future==0.18.3
gast==0.4.0
google-auth==2.23.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.59.0
h11==0.12.0
h2==4.1.0
h5py==3.10.0
hpack==4.0.0
httpcore==0.13.7
httpx==0.19.0
hyperframe==6.0.1
idna==3.4
imageio==2.31.6
imageio-ffmpeg==0.4.9
importlib-metadata==6.8.0
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
langcodes==3.3.0
libclang==16.0.6
librosa==0.8.1
llvmlite==0.38.1
Markdown==3.5
MarkupSafe==2.1.3
moviepy==1.0.3
multidict==6.0.4
murmurhash==1.0.10
norbert==0.2.1
numba==0.55.2
numpy==1.22.4
oauthlib==3.2.2
openai==0.28.1
opt-einsum==3.3.0
packaging==23.2
pandas==1.5.3
Pillow==10.0.1
platformdirs==3.11.0
pooch==1.7.0
preshed==3.0.9
proglog==0.1.10
protobuf==3.20.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pysrt==1.1.2
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
resampy==0.4.2
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.3.1
scipy==1.10.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.0
soundfile==0.12.1
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spleeter==2.3.2
srsly==2.4.8
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-macos==2.13.0
termcolor==2.3.0
thinc==8.2.1
threadpoolctl==3.2.0
tqdm==4.66.1
typer==0.3.2
typing_extensions==4.8.0
urllib3==2.0.7
wasabi==1.1.2
weasel==0.3.3
Werkzeug==3.0.0
wrapt==1.15.0
yarl==1.9.2
zipp==3.17.0
